+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Increment the:
 
 ## [Unreleased]
 
+* [ETW] Fix infinite loop in ETWProvider::close
+  [#3827](https://github.com/open-telemetry/opentelemetry-cpp/pull/3827)
+
 * [CONFIGURATION] File configuration - remove zipkin
   [#3804](https://github.com/open-telemetry/opentelemetry-cpp/pull/3804)
 
@@ -29,6 +32,15 @@ Increment the:
 
 * [SEMANTIC CONVENTIONS] Upgrade to semantic conventions 1.39.0
   [#3813](https://github.com/open-telemetry/opentelemetry-cpp/pull/3813)
+
+* [CONFIGURATION] File configuration - prometheus without_target_info
+  [#3818](https://github.com/open-telemetry/opentelemetry-cpp/pull/3818)
+
+* [BAZEL] Fix CI after bazel 9.0.0 release
+  [#3823](https://github.com/open-telemetry/opentelemetry-cpp/pull/3823)
+
+* [CONFIGURATION] File configuration - implement cardinality_limits
+  [#3820](https://github.com/open-telemetry/opentelemetry-cpp/pull/3820)
 
 Breaking changes: